Fix gl_PerVertex for ES
glslang does not declare gl_ClipDistance as part of gl_PerVertex for
ES by default. This causes SPIRV-Cross (which does declare ClipDistance)
to complain that gl_PerVertex from the shader is undersized. Change
SPIRV-Cross to not declare ClipDistance as part of gl_PerVertex, but to
define it as the pointer to the word following PointSize.